backend/requirements.txt
fastapi
uvicorn
python-dotenv
cloudscraper
groq
httpx
requests
Pillow
aiohttp
groq
